refactor: make legacy configs possible
