+++ prefixes-shared/build.gradle.kts
@@ -1,3 +1,6 @@
 dependencies {
     api(project(":prefixes-api"))
+    implementation("net.kyori:adventure-text-serializer-gson:4.14.0")
+    implementation("net.kyori:adventure-text-serializer-legacy:4.14.0")
+    compileOnly("com.comphenix.protocol:ProtocolLib:5.1.0")
 }
