+++ prefixes-shared/src/main/kotlin/app/simplecloud/plugin/prefixes/shared/MiniMessageImpl.kt
@@ -8,14 +8,23 @@ class MiniMessageImpl {
     companion object {
         private val miniMessage = MiniMessage.miniMessage()
         fun parse(text: String): Component {
+            if(text.contains("§")) {
+                return ComponentSerializerImpl.deserializeLegacy(text)
+            }
             return miniMessage.deserialize(text)
         }
 
         fun parse(text: String, vararg tags: TagResolver): Component {
+            if(text.contains("§")) {
+                return ComponentSerializerImpl.deserializeLegacy(text)
+            }
             return miniMessage.deserialize(text, TagResolver.resolver(tags.asIterable()))
         }
 
         fun parse(text: String, tags: Iterable<TagResolver>): Component {
+            if(text.contains("§")) {
+                return ComponentSerializerImpl.deserializeLegacy(text)
+            }
             return miniMessage.deserialize(text, TagResolver.resolver(tags))
         }
     }
