+++ prefixes-spigot/build.gradle.kts
@@ -1,7 +1,6 @@
 dependencies {
     api(project(":prefixes-shared"))
     compileOnly("org.spigotmc:spigot-api:1.20.4-R0.1-SNAPSHOT")
-    implementation("net.kyori:adventure-text-serializer-gson:4.14.0")
     compileOnly("com.comphenix.protocol:ProtocolLib:5.1.0")
     compileOnly("io.papermc.paper:paper-api:1.20.2-R0.1-SNAPSHOT")
 }
